fix issue that allow decimal for Item Quantity, Cartons and Cube fix display issue when view Order Detail (read-only)

+++ ADJ-Internship/BusinessService/Dtos/ShipmentBookingDtos.cs
@@ -62,7 +62,7 @@ namespace ADJ.BusinessService.Dtos
     //DropList of Boxed/Carton
     public string PackType { get; set; }
 
-    public decimal Quantity { get; set; }
+    public float Quantity { get; set; }
 
     public float Cartons { get; set; }
 
@@ -125,7 +125,7 @@ namespace ADJ.BusinessService.Dtos
     public string Vendor { get; set; }
 
     [Display(Name = "PO Quantity")]
-    public decimal Quantity { get; set; }
+    public float Quantity { get; set; }
 
     //equal Revise Quantity
     [Display(Name = "Booking Quantity")]
